+++ src/datetime.ts
@@ -24,7 +24,7 @@ import {
 import { normalizeZone } from "./impl/zoneUtil";
 import { diff } from "./impl/diff";
 import { parseRFC2822Date, parseISODate, parseHTTPDate, parseSQL } from "./impl/regexParser";
-import { parseFromTokens, explainFromTokens } from "./impl/tokenParser";
+import { parseFromTokens, explainFromTokens, formatOptsToTokens } from "./impl/tokenParser";
 import {
     gregorianToWeek,
     weekToGregorian,
@@ -1234,7 +1234,7 @@ export class DateTime {
 
     /**
      * Create an invalid DateTime.
-     * @param {string} reason - simple string of why this DateTime is invalid. Should not contain parameters or anything else data-dependent
+     * @param {DateTime} reason - simple string of why this DateTime is invalid. Should not contain parameters or anything else data-dependent
      * @param {string} [explanation=null] - longer explanation, may include parameters and other useful debugging information
      * @return {DateTime}
      */
@@ -1256,6 +1256,18 @@ export class DateTime {
      */
     static isDateTime(o: unknown): o is DateTime {
         return !!(o && (o as DateTime)._isLuxonDateTime);
+    }
+
+    /**
+     * Produce the format string for a set of options
+     * @param formatOpts
+     * @param localeOpts
+     * @returns {string}
+     */
+    static parseFormatForOpts(formatOpts: Intl.DateTimeFormatOptions, localeOpts: LocaleOptions = {}) {
+        const tokenList = formatOptsToTokens(formatOpts, Locale.fromObject(localeOpts));
+
+        return !tokenList ? null : tokenList.map((t) => (t ? t.val : null)).join("");
     }
 
     /**